+++ vrp-core/src/utils/mod.rs
@@ -3,8 +3,5 @@
 // Reimport rosomaxa utils
 pub use rosomaxa::utils::*;
 
-mod comparison;
-pub use self::comparison::*;
-
 mod types;
 pub use self::types::Either;